fix(next): isomorphic toolkit catalog — client Connect card must not pull the Composio server module; add fresh-app e2e evidence
Found by the fresh-app install proof: connect-node imported
DEFAULT_INTEGRATION_CATALOG from the server integrations module, dragging
@composio/core (node:module) into the client bundle and 500ing the app.
